Added transformations, moved Quaternions to their own class
